#  Speed-up tests that rely on the database fixture