The `database` and `mutable_database` fixtures were installing and uninstalling the same specs multiple times to maintain the test database in a consistent state. This PR optimizes the procedure by:
- Installing specs only the first time the fixtures are requested
- Caching the installation path into a persistent folder
- Erasing the installation path and copying the cache back into it when needed

The results on my laptop:
```console
$ uname -a
Linux nuvolari 4.15.0-54-generic #58-Ubuntu SMP Mon Jun 24 10:55:24 UTC 2019 x86_64 x86_64 x86_64 GNU/Linux
```
 are the following. On `develop` (with 2 tests not marked):
```console
$ python3.7 bin/spack test -m db
============================================================== test session starts ===============================================================
platform linux -- Python 3.7.4, pytest-3.2.5, py-1.4.34, pluggy-0.4.0
rootdir: /home/mculpo/PycharmProjects/spack/lib/spack/spack/test, inifile: pytest.ini
collected 1553 items                                                                                                                              

database.py .......................
spec_syntax.py .......
cmd/dependencies.py ..
cmd/dependents.py ..
cmd/find.py ....
cmd/graph.py .....
cmd/location.py .......
cmd/module.py ............................
cmd/uninstall.py ...
modules/tcl.py ..

=========================================================== slowest 20 test durations ============================================================
4.01s setup    cmd/location.py::test_location_install_dir
4.00s setup    cmd/module.py::test_exit_with_failure[dotkit-failure_args0]
3.88s setup    cmd/graph.py::test_graph_ascii
3.85s setup    cmd/dependencies.py::test_immediate_installed_dependencies
3.82s setup    cmd/dependents.py::test_immediate_installed_dependents
3.82s setup    cmd/find.py::test_tag1
3.81s teardown database.py::test_old_external_entries_prefix
3.81s setup    spec_syntax.py::TestSpecSyntax::test_spec_by_hash
3.80s teardown database.py::test_regression_issue_8036
3.76s teardown database.py::test_030_db_sanity_from_another_process
3.75s teardown database.py::test_115_reindex_with_packages_not_in_repo
3.33s setup    modules/tcl.py::TestTcl::test_blacklist_implicits
3.22s teardown cmd/module.py::test_setdefault_command
3.21s setup    cmd/uninstall.py::test_multiple_matches
2.59s setup    database.py::test_default_queries
0.79s call     cmd/module.py::test_setdefault_command
0.63s call     database.py::test_025_reindex
0.55s call     cmd/module.py::test_remove_and_add[dotkit]
0.51s call     database.py::test_115_reindex_with_packages_not_in_repo
0.50s call     spec_syntax.py::TestSpecSyntax::test_spec_by_hash
============================================================= 1470 tests deselected ==============================================================
================================================== 83 passed, 1470 deselected in 64.46 seconds ===================================================
```
while with this PR:
```console
$ python3.7 bin/spack test -m db
============================================================== test session starts ===============================================================
platform linux -- Python 3.7.4, pytest-3.2.5, py-1.4.34, pluggy-0.4.0
rootdir: /home/mculpo/PycharmProjects/spack/lib/spack/spack/test, inifile: pytest.ini
collected 1553 items                                                                                                                              

database.py .......................
spec_syntax.py .......
cmd/buildcache.py .
cmd/debug.py .
cmd/dependencies.py ..
cmd/dependents.py ..
cmd/find.py ....
cmd/graph.py .....
cmd/location.py .......
cmd/module.py ............................
cmd/uninstall.py ...
modules/tcl.py ..

=========================================================== slowest 20 test durations ============================================================
2.45s setup    database.py::test_default_queries
0.74s call     cmd/module.py::test_setdefault_command
0.65s call     database.py::test_025_reindex
0.56s call     cmd/buildcache.py::test_buildcache_preview_just_runs
0.55s call     cmd/module.py::test_remove_and_add[dotkit]
0.54s call     cmd/module.py::test_remove_and_add[tcl]
0.54s call     spec_syntax.py::TestSpecSyntax::test_spec_by_hash
0.52s call     database.py::test_115_reindex_with_packages_not_in_repo
0.45s call     modules/tcl.py::TestTcl::test_autoload_with_constraints
0.28s call     cmd/uninstall.py::test_recursive_uninstall
0.25s call     database.py::test_060_remove_and_add_root_package
0.23s call     database.py::test_030_db_sanity_from_another_process
0.22s call     database.py::test_070_remove_and_add_dependency_package
0.21s call     database.py::test_010_all_install_sanity
0.20s call     database.py::test_090_non_root_ref_counts
0.19s call     database.py::test_020_db_sanity
0.15s call     database.py::test_015_write_and_read
0.15s call     database.py::test_080_root_ref_counts
0.14s call     cmd/module.py::test_find_recursive
0.13s call     cmd/dependents.py::test_transitive_installed_dependents
============================================================= 1468 tests deselected ==============================================================
================================================== 85 passed, 1468 deselected in 15.72 seconds ===================================================
```