@alalazo 2019-07-15
On Travis it seems we gain 2-3 mins per unit test run.