@alalazo 2019-07-16
> If we use these fixtures in the right places, the database setup and teardown only need to happen once per session.

I also thought the same initially, but this is not true for a couple of reasons:

1. Each test using the function scoped fixture `mutable_database` needs to restore the state of the database during the teardown. Doing it with uninstall and install commands takes much longer than copying a cached version of the store in place.
2. The `database` fixture also plugs the test store and DB into Spack and depends on other two fixtures (`mock_packages` and `config`) that do the same for the `builtin.mock` repository and the mock configuration. All these fixtures are module scoped and cannot be simply changed to be session scoped (otherwise all the tests will operate on mock packages with mock configs and mock store).

These considerations led me to think that the best way to speed-up this part of tests was to construct the mock database on first request and store it in a place that is available for the whole test session.

> I like the fact that this code block actually tests uninstall_by_spec() at the end of testing. Do we lose coverage by removing it?

I restarted the tests after rebasing on #12032  to see if we are losing coverage. Regardless of coverage, if we are not testing `uninstall_by_spec()` I'd prefer to add a new test for it rather than relying on it being executed as part of a fixture's teardown.