@tgamblin 2019-07-16
@alalazo: ok all that sounds good, but can `database` not still be session-scoped instead of module-scoped?