@alalazo 2019-07-16
> ok all that sounds good, but can database not still be session-scoped instead of module-scoped?

If we really want that I think we need to replicate much of what `mock_packages` and `config` fixtures do in the setup of `database` as a session scoped fixture can't depend on module scoped fixtures.