@tgamblin 2019-07-16
> If we really want that I think we need to replicate much of what mock_packages and config fixtures do in the setup of database as a session scoped fixture can't depend on module scoped fixtures.

Can't these also just be moved to session scope?  They are mostly used in immutable contexts, though we may have to make mutable variants of them.  I think *most* of the code that messes with them restores them after each function, though.