@tldahlgren 2019-07-16
> I restarted the tests after rebasing on #12032 to see if we are losing coverage. Regardless of coverage, if we are not testing `uninstall_by_spec()` I'd prefer to add a new test for it rather than relying on it being executed as part of a fixture's teardown.

FYI.  There are new tests for `uninstall_by_spec` in https://github.com/spack/spack/pull/11971, albeit for the error paths.