@alalazo 2019-07-18
> Can't these also just be moved to session scope? They are mostly used in immutable contexts, though we may have to make mutable variants of them. I think most of the code that messes with them restores them after each function, though.

Just checked and by just moving them to session scope we have tests failures, for tests that either need the actual Spack's configuration to be plugged in (e.g. `test_all_virtual_packages_have_default_providers`) or for tests that need to operate on the builtin repository (e.g. `test_all_versions_are_lowercase`). 

We can invert the logic and assume that, if not stated otherwise, a test in Spack uses the mock repository and config - but I don't know how much we need to refactor for that. Should I have a look or are we keeping this for another PR?