@alalazo 2019-07-20
@tgamblin 63a92d7a28fab998001cc3368b5e1c1153e1a620 constructs the database once per session and avoids needless copying:

https://github.com/spack/spack/blob/63a92d7a28fab998001cc3368b5e1c1153e1a620/lib/spack/spack/test/conftest.py#L431-L434

I think that's the best I can do right now. During the process I also enforced the DB returned by that fixture to be read-only:

https://github.com/spack/spack/blob/63a92d7a28fab998001cc3368b5e1c1153e1a620/lib/spack/spack/test/conftest.py#L436-L437

and discovered several tests that were claiming not to change the DB but failed to fulfill the promise. They have now been changed to use `mutable_database` instead. Let me know if you have comments or requests on the latest modifications in the PR.