@tgamblin 2019-07-20
@alalazo: LGTM and thanks for all the work on this.  This speeds the DB tests up ~4x on my laptop (takes 70 seconds of the test time).  It'd be nice to eventually do this using session scope but this is already a really big improvement.